+++ src/QueryCat.Backend/Commands/Select/SelectPlanner.Pipeline.cs
@@ -285,8 +285,8 @@ internal sealed partial class SelectPlanner
         if (outputIterator.HasOutputDefined)
         {
             context.HasOutput = true;
-            var resultIterator = !hasVaryingTarget
-                ? new AdjustColumnsLengthsIterator(outputIterator)
+            var resultIterator = !hasVaryingTarget && ExecutionThread.Options.AnalyzeRowsCount > 0
+                ? new AdjustColumnsLengthsIterator(outputIterator, ExecutionThread.Options.AnalyzeRowsCount)
                 : (IRowsIterator)outputIterator;
             var actionIterator = new ActionRowsIterator(resultIterator, "write to output")
             {