+++ src/QueryCat.Backend/Execution/ExecutionOptions.cs
@@ -77,6 +77,11 @@ public sealed class ExecutionOptions
     /// </summary>
     public bool RunBootstrapScript { get; set; } = true;
 
+    /// <summary>
+    /// Number of rows to analyze. This affects columns types detection and column length adjustment. -1 to analyze all.
+    /// </summary>
+    public int AnalyzeRowsCount { get; set; } = 10;
+
     public ExecutionOptions(TextTableOutput.Style outputStyle = TextTableOutput.Style.Table)
     {
         var tableOutput = new TextTableOutput(