+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add string functions "starts_with", "split_part", "regexp_count", "regexp_substr".
 - Allow to read/write from/to gzip archives.
 - Add "VALUES", "AT TIME ZONE" clauses.
+- Add option "analyze-rows".
 
 ### Changed